Implement RwCAQ separately from GsQuery.
Started implementing RwCAQ without using GsQuery implementation. Ideally, RwCAQ should be usable outside of GemStone. PetiteParser was chosen because it is provided in a base GemStone extent and it is also available for other Smalltalk variants.

The design is also changing. The conditional expression on a Component will be parsed, then evaluated with a context. The result of evaluation will be a boolean value. If it evaluates to true, the component should be loaded. If it evaluates to false, the component should not be loaded. Should parsing or evaluation fail, an exception will be raised.